Fix #17 cleanup unused API
- Renaming RoundEnvironementInternal
- Refactor Context and Context internal
 - remove old method
 - add objectGraph
