+++ core/src/main/java/io/nuun/kernel/core/internal/KernelGuiceModuleInternal.java
@@ -25,8 +25,6 @@ import com.google.inject.util.Providers;
 import io.nuun.kernel.api.di.UnitModule;
 import io.nuun.kernel.api.plugin.context.Context;
 import io.nuun.kernel.core.KernelException;
-import io.nuun.kernel.core.internal.context.ContextInternal;
-import io.nuun.kernel.core.internal.context.InitContextInternal;
 import io.nuun.kernel.spi.Concern;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
