--- specs/src/main/java/io/nuun/kernel/api/plugin/RoundEnvironment.java
@@ -1,37 +1,0 @@
-/**
- * Copyright (C) 2013-2014 Kametic <epo.jemba@kametic.com>
- *
- * Licensed under the GNU LESSER GENERAL PUBLIC LICENSE, Version 3, 29 June 2007;
- * or any later version
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.gnu.org/licenses/lgpl-3.0.txt
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.nuun.kernel.api.plugin;
-
-/**
- * The kernel initializes all the plugins inside rounds.
- * 
- * @author epo.jemba{@literal @}kametic.com
- */
-public interface RoundEnvironment
-{
-
-    /**
-     * @return the current round number.
-     */
-    public int roundNumber();
-
-    /**
-     * @return true if we are in the first round.
-     */
-    public boolean firstRound();
-    
-}
